leave broken vscode behavior
